+++ Changelog.md
@@ -2,6 +2,21 @@
 
 This page contains information about changes to the PowerBI Visual Tools (pbiviz).
 
+## 6.1.2
+* Fixed issue with `process/browser` not being resolved correctly in webpack config
+
+## 6.1.1
+* Fixed bug with source maps generation (now using `eval-source-map` instead of `source-map`)
+
+## 6.1.0
+* Added `--certification-audit` flag to `pbiviz package` command. It audits the visual package for possible certification issues.
+* Added `--certification-fix` flag to `pbiviz package` command. It fixes found certification issues in the visual package. Please carefully test the built package to ensure its functionality is not broken after that.
+
+## 6.0.0
+* **Introduced pbiviz.js support.** This feature allows generate configuration dynamically using benefits of JavaScript.
+  It allows you to easily use any js logic to build your pbiviz file using environment variables or other parameters.
+* Updated dependencies to the latest versions.
+
 ## 5.6.0
 * Updated to ESLint v9
 * Updated to TypeScript v5